requirements.txt
PySide6
requests
pytz
QT-PyQt-PySide-Custom-Widgets
firebase-admin